+++ src/components/iRepeat.js
@@ -1,13 +1,13 @@
 // Use performance.now() to control the "FPS" of a repeatedly called function
 
-export function iRepeat(fun, fps = 60) {
+export function iRepeat(fun, options = { framerate: 60, break: false }) {
   let now = 0,
     then = 0,
     elapsed = 0
-  const interval = 1000 / fps
+  const interval = 1000 / options.framerate
 
   const r = () => {
-    requestAnimationFrame(r)
+    if (!options.break) requestAnimationFrame(r)
 
     now = performance.now()
     elapsed = now - then